+++ tests/async/test_calendar_getters.py
@@ -50,9 +50,7 @@ async def client():
 async def test_calendar(client, args, kwargs, expected):
     ts = await client.get_calendar(*args, **kwargs)
     assert isinstance(ts, expected)
-    assert isinstance(
-        expected == list and ts[0] or ts["1"][0], aladhan.Timings
-    )
+    assert isinstance(expected == list and ts[0] or ts["1"][0], aladhan.Timings)
 
 
 @pytest.mark.asyncio
@@ -91,9 +89,7 @@ async def test_calendar(client, args, kwargs, expected):
 async def test_calendar_by_address(client, args, kwargs, expected):
     ts = await client.get_calendar_by_address(*args, **kwargs)
     assert isinstance(ts, expected)
-    assert isinstance(
-        expected == list and ts[0] or ts["1"][0], aladhan.Timings
-    )
+    assert isinstance(expected == list and ts[0] or ts["1"][0], aladhan.Timings)
 
 
 @pytest.mark.asyncio
@@ -147,6 +143,4 @@ async def test_calendar_by_address(client, args, kwargs, expected):
 async def test_calendar_by_city(client, args, kwargs, expected):
     ts = await client.get_calendar_by_city(*args, **kwargs)
     assert isinstance(ts, expected)
-    assert isinstance(
-        expected == list and ts[0] or ts["1"][0], aladhan.Timings
-    )
+    assert isinstance(expected == list and ts[0] or ts["1"][0], aladhan.Timings)