+++ docs/changelog.rst
@@ -30,10 +30,13 @@ v1.2.0 (Unreleased)
 - :class:`~aladhan.enums.Shafaq`
 - Added a ``shafaq`` parameter for :class:`Parameters`
 - :exc:`~aladhan.exceptions.InvalidShafaq`
+- Adding the new changes to the api.
+    - :attr:`Timings.first_third`
+    - :attr:`Timings.last_third`
 
 **Changed**
 
-- Fixed date converters, they were switched around :l
+- Fixed date converters, they were switched around ...
 
 v1.1.0
 ------